[SCB-1862] ServiceComb add extensions to support cas env variables:fixed as reviewed
