+++ dependencies/bom/pom.xml
@@ -287,6 +287,12 @@
         <artifactId>java-chassis-spring-boot-starter-standalone</artifactId>
         <version>${project.version}</version>
       </dependency>
+      <!-- ServiceComb ServiceStage extension -->
+      <dependency>
+        <groupId>org.apache.servicecomb</groupId>
+        <artifactId>servicestage</artifactId>
+        <version>${project.version}</version>
+      </dependency>
       <!-- ServiceComb: swagger -->
       <dependency>
         <groupId>org.apache.servicecomb</groupId>
@@ -363,12 +369,6 @@
       <dependency>
         <groupId>org.apache.servicecomb</groupId>
         <artifactId>transport-rest-vertx</artifactId>
-        <version>${project.version}</version>
-      </dependency>
-      <!-- ServiceComb third-extensions -->
-      <dependency>
-        <groupId>org.apache.servicecomb</groupId>
-        <artifactId>cas-env-variables</artifactId>
         <version>${project.version}</version>
       </dependency>
     </dependencies>
