+++ huawei-cloud/servicestage/src/main/resources/META-INF/services/org.apache.servicecomb.serviceregistry.adapter.EnvAdapter
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.servicecomb.cas.CasEnvVariablesAdapter
+org.apache.servicecomb.huaweicloud.servicestage.CasEnvVariablesAdapter
